refactor(linter/vue): extract computed getter utils into utils/vue.rs
